Refactor wram stuff out of main
