+++ src/main.c
@@ -8,6 +8,7 @@
 #include "lib/unrle.h"
 #include "mmc3/mmc3_code.h"
 #include "mmc3/mmc3_code.c"
+#include "wram.h"
 #include "nametable_loader.h"
 #include "dungeon.h"
 #include "players.h"
@@ -85,32 +86,6 @@ unsigned char current_screen;
 
 char irq_array[32];
 unsigned char double_buffer[32];
-
-#pragma bss-name(push, "XRAM")
-// extra RAM at $6000-$7fff
-
-#define WRAM_VERSION 0x0004
-unsigned int wram_start;
-unsigned char dungeon_layout_initialized;
-unsigned char wram_dungeon_layout[NUM_DUNGEONS * NUM_DUNGEON_LEVELS];
-unsigned char party_initialized;
-unsigned char player_name[4][5];
-unsigned char player_str[4];
-unsigned char player_int[4];
-unsigned char player_wis[4];
-unsigned char player_dex[4];
-unsigned char player_con[4];
-unsigned int player_hp[4];
-unsigned int player_max_hp[4];
-unsigned int player_mp[4];
-unsigned int player_max_mp[4];
-unsigned int player_xp[4];
-unsigned int player_lv[4];
-player_class_type player_class[4];
-
-unsigned char unrle_buffer[1024];
-
-#pragma bss-name(pop)
 
 // the fixed bank
 
@@ -603,12 +578,6 @@ void start_game (void) {
   reset_cursor();
 
   current_screen = 0;
-}
-
-void init_wram (void) {
-  if (wram_start != WRAM_VERSION)
-    memfill(&wram_start,0,0x2000);
-  wram_start = WRAM_VERSION;
 }
 
 void flip_screen (void) {
