+++ src/wram.h
@@ -1,0 +1,6 @@
+extern unsigned char unrle_buffer[];
+extern unsigned char dungeon_layout_initialized;
+extern unsigned char party_initialized;
+extern unsigned char wram_dungeon_layout[];
+
+void init_wram (void);
